+++ src/main/scala/edu/ie3/simona/ontology/messages/ServiceMessage.scala
@@ -8,7 +8,7 @@ package edu.ie3.simona.ontology.messages
 
 import edu.ie3.simona.agent.participant.ParticipantAgent
 import edu.ie3.simona.agent.participant.ParticipantAgent.ParticipantRequest
-import edu.ie3.simona.api.data.ontology.DataMessageFromExt
+import edu.ie3.simona.api.ontology.DataMessageFromExt
 import edu.ie3.simona.model.participant.evcs.EvModelWrapper
 import edu.ie3.simona.scheduler.ScheduleLock.ScheduleKey
 import edu.ie3.simona.service.ServiceStateData.InitializeServiceStateData
@@ -45,9 +45,9 @@ object ServiceMessage {
     * @param data
     *   The data, that is used during the registration.
     */
-  final case class SecondaryServiceRegistrationMessage[D](
+  final case class SecondaryServiceRegistrationMessage(
       requestingActor: ActorRef[ParticipantAgent.Request],
-      data: D,
+      data: Any,
   ) extends ServiceRegistrationMessage
 
   /** Message to register with a primary data service.