+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Authors.md file [#1503](https://github.com/ie3-institute/simona/issues/1503)
 - Quantity to squants conversion in WeatherService [#1506](https://github.com/ie3-institute/simona/issues/1506)
 - Considering primary data that start before simulation [#1034](https://github.com/ie3-institute/simona/issues/1034)
+- Implement time series retrieval for WeatherSource [#1511](https://github.com/ie3-institute/simona/issues/1511)
+- Implement weather forecast provision by WeatherService [#1512](https://github.com/ie3-institute/simona/issues/1512)
+- Introduce optimizing EM strategy [#1500](https://github.com/ie3-institute/simona/issues/1500)
+- Introduce `DomesticHotWaterStorage` [#1481](https://github.com/ie3-institute/simona/issues/1481)
 - Adding transformer tap changing to congestion mitigation [#1378](https://github.com/ie3-institute/simona/issues/1378)
 
 ### Changed
@@ -40,6 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Simplify flex option result configuration [#1111](https://github.com/ie3-institute/simona/issues/1111)
 - Harmonisation of `TestInputData` for `Transformer` [#163](https://github.com/ie3-institute/simona/issues/163)
 - Refactored method to determine next Threshold within `ThermalGrid` [#1477](https://github.com/ie3-institute/simona/issues/1477)
+- Unify coordinate class [#1516](https://github.com/ie3-institute/simona/issues/1516)
+- Remove type parameters from data and message classes [#1524](https://github.com/ie3-institute/simona/issues/1524)
+- Adapt ThermalHouse and HP flexibility behaviour [#1391](https://github.com/ie3-institute/simona/issues/1391)
 
 ### Fixed
 - Fixes in Documentation, ScalaDocs, Code Style and more [#1397](https://github.com/ie3-institute/simona/issues/1397)
@@ -53,11 +60,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed CFF-Version [#1464](https://github.com/ie3-institute/simona/issues/1464)
 - Fixed handleInput of `ProfileLoadModel` [#1441](https://github.com/ie3-institute/simona/issues/1441)
 - Fixed compiler warnings related to import and scala3 syntax [#1383](https://github.com/ie3-institute/simona/issues/1383)
+- Cleaned up dependencies [#1539](https://github.com/ie3-institute/simona/issues/1539)
 
 ### Removed
 - Removed unused classes and methods related to pekko classic actors [#1389](https://github.com/ie3-institute/simona/issues/1389)
 - Removed dedicated `random` load model behaviour [#1436](https://github.com/ie3-institute/simona/issues/1436)
 - Removed some units from scalaDoc within `ThermalHouse` [#1495](https://github.com/ie3-institute/simona/issues/1495)
+- Removed unused code in test directory [#1533](https://github.com/ie3-institute/simona/issues/1533)
+
+### Updates
+- Updated gradle to version 9.0.0 [#1528](https://github.com/ie3-institute/simona/issues/1528)
 
 ## [4.0.0] - 2025-05-09